preview server: extract RouteBuilder
