+++ preview/src/main/scala/laika/preview/ServerBuilder.scala
@@ -16,11 +16,8 @@
 
 package laika.preview
 
-import java.io.{File, InputStream}
-
 import cats.syntax.all._
 import cats.effect._
-import fs2.io.readInputStream
 import laika.api.Renderer
 import laika.api.builder.OperationConfig
 import laika.format.HTML
@@ -29,10 +26,8 @@ import laika.io.implicits._
 import laika.io.api.{TreeParser, TreeRenderer}
 import laika.io.model.{InputTreeBuilder, StringTreeOutput}
 import laika.theme.ThemeProvider
-import org.http4s.EntityEncoder.entityBodyEncoder
 import org.http4s._
 import org.http4s.dsl.Http4sDsl
-import org.http4s.headers.`Content-Type`
 import org.http4s.implicits._
 import org.http4s.server.{Router, Server}
 import org.http4s.server.blaze.BlazeServerBuilder
@@ -48,17 +43,7 @@ class ServerBuilder[F[_]: Async] (parser: Resource[F, TreeParser[F]],
                                   port: Int,
                                   pollInterval: FiniteDuration) extends Http4sDsl[F] {
 
-  implicit def inputStreamResourceEncoder[G[_]: Sync, IS <: InputStream]: EntityEncoder[G, Resource[G, IS]] =
-    entityBodyEncoder[G].contramap { (in: Resource[G, IS]) =>
-      fs2.Stream.resource(in).flatMap { stream =>
-        readInputStream[G](Sync[G].pure(stream), 4096, closeAfterUse = false)
-      }
-    }
-
   def build: Resource[F, Server] = {
-    
-    def mediaTypeFor (suffix: String): Option[MediaType] =
-      MediaType.forExtension(suffix).orElse(MediaType.forExtension(suffix.split("\\.").last))
     
     def htmlRenderer (config: OperationConfig): Resource[F, TreeRenderer[F]] = Renderer
       .of(HTML)
@@ -108,25 +93,7 @@ class ServerBuilder[F[_]: Async] (parser: Resource[F, TreeParser[F]],
           parser.config.docTypeMatcher
         ).flatMap { _ =>
 
-          val routes = HttpRoutes.of[F] {
-
-            case GET -> path =>
-              val laikaPath = laika.ast.Path.parse(path.toString)
-              cache.get.map(_.get(laikaPath.withoutFragment)).flatMap {
-                case Some(Right(content)) =>
-                  println(s"serving path $laikaPath - transformed markup")
-                  Ok(content).map(_.withContentType(`Content-Type`(MediaType.text.html)))
-                case Some(Left(input)) =>
-                  println(s"serving path $laikaPath - static input")
-                  val mediaType = laikaPath.suffix.flatMap(mediaTypeFor).map(`Content-Type`(_))
-                  Ok(input).map(_.withHeaders(Headers(mediaType)))
-                case None =>
-                  println(s"serving path $laikaPath - not found")
-                  NotFound()
-              }
-          }
-
-          val httpApp = Router("/" -> routes).orNotFound
+          val httpApp = Router("/" -> new RouteBuilder[F](cache).build).orNotFound
 
           BlazeServerBuilder[F](ctx)
             .bindHttp(port, "localhost")
